+++ packages/zeebe-element-templates-json-schema/test/fixtures/invalid-binding-type.js
@@ -50,14 +50,15 @@ export const errors = [
               'zeebe:calledElement',
               'zeebe:linkedResource',
               'zeebe:userTask',
-              'zeebe:formDefinition'
+              'zeebe:formDefinition',
+              'zeebe:calledDecision'
             ]
           },
           message: 'should be equal to one of the allowed values'
         }
       ]
     },
-    message: 'invalid property.binding type "foo"; must be any of { property, zeebe:taskDefinition:type, zeebe:input, zeebe:output, zeebe:property, zeebe:taskHeader, bpmn:Message#property, bpmn:Message#zeebe:subscription#property, zeebe:taskDefinition, zeebe:calledElement, zeebe:linkedResource, zeebe:userTask }'
+    message: 'invalid property.binding type "foo"; must be any of { property, zeebe:taskDefinition:type, zeebe:input, zeebe:output, zeebe:property, zeebe:taskHeader, bpmn:Message#property, bpmn:Message#zeebe:subscription#property, zeebe:taskDefinition, zeebe:calledElement, zeebe:linkedResource, zeebe:userTask, zeebe:formDefinition, zeebe:calledDecision }'
   },
   {
     dataPath: '',